Move HtmlSantiazer into dom package
