+++ packages/roosterjs-editor-plugins/lib/Paste/Paste.ts
@@ -2,15 +2,9 @@ import buildClipboardData from './buildClipboardData';
 import fragmentHandler from './fragmentHandler';
 import textToHtml from './textToHtml';
 import { Editor, EditorPlugin } from 'roosterjs-editor-core';
-import { getFormatState } from 'roosterjs-editor-api';
-import { insertImage } from 'roosterjs-editor-api';
+import { getFormatState, insertImage } from 'roosterjs-editor-api';
 import {
     AttributeCallbackMap,
-    getInheritableStyles,
-    HtmlSanitizer,
-    htmlToDom,
-} from 'roosterjs-html-sanitizer';
-import {
     BeforePasteEvent,
     ChangeSource,
     ClipboardData,
@@ -23,7 +17,10 @@ import {
     applyFormat,
     fromHtml,
     getFirstLeafNode,
+    getInheritableStyles,
     getNextLeafSibling,
+    HtmlSanitizer,
+    htmlToDom,
     Position,
 } from 'roosterjs-editor-dom';
