+++ packages/roosterjs/lib/index.ts
@@ -5,5 +5,4 @@ export * from 'roosterjs-editor-core';
 export * from 'roosterjs-editor-api';
 export * from 'roosterjs-editor-plugins';
 export * from 'roosterjs-plugin-image-resize';
-export * from 'roosterjs-html-sanitizer';
 export * from 'roosterjs-plugin-picker';
